internal/browser/pyproject/requirements.txt
Django==2.2.2
django-rest-framework==0.1.0
djangorestframework==3.9.4
django-rest-framework==0.1.0
django-filter==2.1.0
djangorestframework-jwt==1.11.0
django-enumfields==1.0.0
django-extensions==2.1.9
psycopg2-binary==2.8.4
requests==2.22.0
model-mommy==1.6.0
uuid==1.30
coverage==4.5.3
celery==4.4.0
redis==3.2.1
django-redis==4.11.0
Pillow==6.1.0
django-storages==1.7.1
elasticsearch-dsl==7.0.0
